WIP: fix unnecessary typecasts in Rust (#2366)
Note: includes https://github.com/sourmash-bio/sourmash/pull/2364

+++ src/core/src/encodings.rs
@@ -352,9 +352,9 @@ pub fn to_aa(seq: &[u8], dayhoff: bool, hp: bool) -> Result<Vec<u8>, Error> {
 
         let residue = translate_codon(chunk)?;
         if dayhoff {
-            converted.push(aa_to_dayhoff(residue) as u8);
+            converted.push(aa_to_dayhoff(residue));
         } else if hp {
-            converted.push(aa_to_hp(residue) as u8);
+            converted.push(aa_to_hp(residue));
         } else {
             converted.push(residue);
         }
